refactor: remove sql package
