+++ errors.go
@@ -1,4 +1,4 @@
-package request
+package greptime
 
 import (
 	"errors"
@@ -10,6 +10,7 @@ var (
 	ErrEmptyTimestamp       = errors.New("timestamp should not be empty")
 	ErrEmptyKey             = errors.New("key should not be empty")
 	ErrEmptyQuery           = errors.New("sql or promql is required in querying")
+	ErrEmptyStep            = errors.New("step is required in promql")
 	ErrInvalidTimePrecision = errors.New("precision of timestamp is not valid")
 	ErrNoSeriesInMetric     = errors.New("empty series in Metric")
 )
