+++ mask.go
@@ -1,4 +1,4 @@
-package request
+package greptime
 
 import (
 	"bytes"
@@ -7,15 +7,19 @@ import (
 	"github.com/bits-and-blooms/bitset"
 )
 
+// Mask is help to set null bits.
 type Mask struct {
 	bs bitset.BitSet
 }
 
+// set is to set which position is to be set to 1
 func (n *Mask) set(idx uint) *Mask {
 	n.bs.Set(idx)
 	return n
 }
 
+// shrink is to help to generate the bytes number the caller is interested
+// via LittleEndian
 func (n *Mask) shrink(bSize int) ([]byte, error) {
 	if n.bs.Len() == 0 {
 		return nil, nil
