--- pkg/request/promql.go
@@ -1,8 +1,0 @@
-package request
-
-type PromQL struct {
-	Query string
-	Start string
-	End   string
-	Step  string
-}
