+++ request_test.go
@@ -1,7 +1,8 @@
-package request
+package greptime
 
 import (
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -17,7 +18,21 @@ func TestQueryBuilder(t *testing.T) {
 	assert.Nil(t, request)
 	assert.ErrorIs(t, err, ErrEmptyQuery)
 
+	// test Sql
 	rb.WithSql("select * from monitor")
+	request, err = rb.Build(&Config{})
+	assert.NotNil(t, request)
+	assert.Nil(t, err)
+
+	// reset Sql to test RangePromql
+	rb.WithSql("")
+	rp := RangePromql{
+		Query: "up == 0",
+		Start: time.Now(),
+		End:   time.Now(),
+		Step:  "10s",
+	}
+	rb.WithRangePromql(rp)
 	request, err = rb.Build(&Config{})
 	assert.NotNil(t, request)
 	assert.Nil(t, err)
@@ -25,13 +40,14 @@ func TestQueryBuilder(t *testing.T) {
 
 func TestInsertBuilder(t *testing.T) {
 	r := InsertRequest{}
+
 	// empty database
 	req, err := r.Build(&Config{})
 	assert.Equal(t, ErrEmptyDatabase, err)
 	assert.Nil(t, req)
 
 	// empty table
-	r.Database = "public"
+	r.header = header{"public"}
 	req, err = r.Build(&Config{})
 	assert.Equal(t, ErrEmptyTable, err)
 	assert.Nil(t, req)
